refactor(harness): apply simplify review to the measurement decomposition
Four deduped fixes from the 4-angle simplify review (no behavior change):

1. Dedupe the harness-sourced batch branch: capacity_batch_size and
   static_batch_size shared a verbatim block; extract
   ExperimentConfig._harness_sourced_batch_size() that both call.

2. Seam honesty: move warmup_result into SourceMetrics so the mode-agnostic
   assemble_experiment_result carries no offline-vocabulary type. The offline
   producer packs it; a source without a warmup phase leaves it None (field
   typed Any | None, documented).

3. Extract the warnings cluster (collect_measurement_warnings wrapper,
   _check_persistence_mode, collect_warnings) out of result_assembly into the
   existing measurement_warnings module (same concern family, no import cycle;
   it runs before assembly with zero SourceMetrics dependency). result_assembly
   now holds just the producer/assembler seam.

4. Rename harness/persistence.py -> harness/staging.py: it stages into a temp
   dir that the S4 BundleWriter finalises; the name now matches the layering.

Patch-target docstrings and test patch strings updated to the new use sites.
